tests/LLM_AGENT/requirements.txt
requests
pyautogui
pillow
pytesseract
pyttsx3
